+++ tests/test_window.py
@@ -66,14 +66,12 @@ def test_imshow_pil():
 
 
 def test_imshow_window_pos():
-    with (
-        cv3.Window('pos 0,0', pos=(0,0)) as w1,
-        cv3.Window('pos 960,540', pos=(480,270)) as w2,
-        cv3.Window('pos 0,700', pos=(0,540)) as w3,
-    ):
-        for w in w1, w2, w3:
-            w.imshow(test_img)
-        cv3.wait_key(2000)
+    with cv3.Window('pos 0,0', pos=(0,0)) as w1:
+        with cv3.Window('pos 960,540', pos=(480,270)) as w2:
+            with cv3.Window('pos 0,700', pos=(0,540)) as w3:
+                for w in (w1, w2, w3):
+                    w.imshow(test_img)
+                cv3.wait_key(2000)
 
 
 def test_window_move():